refactor(map-corridors): share one no-GPS sort comparator across tray and list
The no-GPS sort was effectively duplicated: NoGpsTray used
compareNoGpsPhotos (filename primary, EXIF timestamp tie-break) while
groupPhotosByFlag sorted by compareFilenames alone (no tie-break), so the
tray and the right-side list's no-GPS group could order identical-filename
photos differently.

Move compareNoGpsPhotos into types/markers.ts (next to NoGpsPhoto and
compareFilenames) as the single source of truth; NoGpsTray re-exports it for
back-compat (and the componentLogic test), and groupPhotosByFlag now sorts
the no-GPS group through it. Both surfaces share the exact tie-break order.
